♻️ extract frontend and evaluation from main package including dependencies (#403)
This PR extracts the frontend `mqt.bench.viewer` and the evaluation
functionality `mqt.bench.evaluation` from the main `mqt.bench` package
into separate namespace packages in the top level of the repository.

Since we are planning to move away from the current frontend at some
point, it only makes sense to separate the code from the actual
benchmark generation logic.
Additionally, evaluation code need not necessarily be part of the
package itself. Especially when it comes to results.

This slims down the package slightly and allows to drop some of the
dependencies.

> [!NOTE]
> No code in the respective subpackages has been changed. They are just
packaged separately now (including their own lockfile). They can still
be installed on top of mqt-bench and will continue to work as they have
before this PR once installed properly.

---------

Signed-off-by: burgholzer <burgholzer@me.com>